adding sagecam wiring diagram for plugin, pinpoint, v2
v2 because it's the version with the 3 port active poe switch,
and the poe converter for the nanobeam specifically. its less
parts than v1. but this one is the wall plug in version, and
has the google coral stuff. this was the one deployed at rcc, it's
a bit niche but it does encompass v2 and plug in capabilities